Standardize API patterns across exchanges
- Add Serde.isempty for empty strings to 8 exchanges (Kraken, Deribit,
  Mexc, Huobi, Bitfinex, Bithumb, Upbit, Aevo)
- Fix request_sign! return types in Bitfinex and Huobi (Nothing → Q)
- Fix OKX maxLmtSz/maxMktSz/maxStopSz types (Int64 → Float64)
- Fix MEXC contractAddress type (String → Maybe{String})
- Fix Binance FAPI settlePlan type (Union{Nothing,Int64} → Maybe{Int64})

+++ src/Okx/API/V5/Public/Instruments.jl
@@ -61,12 +61,12 @@ struct InstrumentsData <: OkxData
     lotSz::Maybe{Float64}
     maxIcebergSz::Maybe{Float64}
     maxLmtAmt::Maybe{Float64}
-    maxLmtSz::Maybe{Int64}
+    maxLmtSz::Maybe{Float64}
     maxMktAmt::Maybe{Float64}
-    maxMktSz::Maybe{Int64}
+    maxMktSz::Maybe{Float64}
     maxPlatOICoinLmt::Maybe{String}
     maxPlatOILmt::Maybe{String}
-    maxStopSz::Maybe{Int64}
+    maxStopSz::Maybe{Float64}
     maxTriggerSz::Maybe{Float64}
     maxTwapSz::Maybe{Float64}
     method::Maybe{String}
